Moved key contexts into controller folder
