--- src/edu/uco/sdd/rocketdog/model/IBehavior.java
@@ -1,6 +1,0 @@
-package edu.uco.sdd.rocketdog.model;
-
-interface IBehavior {
-    void update();
-    
-}
